+++ integrations/entity-async-graphql-macros/tests/attribute/mod.rs
@@ -1,0 +1,2 @@
+mod r#enum;
+mod r#struct;