+++ integrations/entity-async-graphql-macros/tests/attribute/struct/mod.rs
@@ -1,0 +1,1 @@
+mod gql_ent;